[#639] ExternalServices 신설 — LinkPreview 엔진 분리, SwiftLinkPreview 의존을 소비 모듈로 이전
— LinkPreviewFetchEngine 프로토콜 신설(PlaceSuggestEngine 동형) — usecase 구현체는 Domain 잔류, SwiftLinkPreview 래핑 엔진(SwiftLinkPreviewBaseFetchEngineImple)만 ExternalServices에
— Common3rdParty에서 SwiftLinkPreview 제거 → Domain 재오염을 컴파일이 차단. 앱은 엔진 직접 생성이 사라져 직접 의존 불필요
— LinkPreviewFetchUsecaesImple 오타 클래스명을 LinkPreviewFetchUsecaseImple로 정정
