+++ Services/ExternalServices/ExternalServices.h
@@ -1,0 +1,16 @@
+//
+//  ExternalServices.h
+//  ExternalServices
+//
+//  Created by sudo.park on 7/30/26.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for ExternalServices.
+FOUNDATION_EXPORT double ExternalServicesVersionNumber;
+
+//! Project version string for ExternalServices.
+FOUNDATION_EXPORT const unsigned char ExternalServicesVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ExternalServices/PublicHeader.h>
