+++ Supports/Common3rdParty/Project.swift
@@ -23,7 +23,6 @@ let project = Project.framework(
         .external(name: "GoogleSignIn"),
         .external(name: "GoogleSignInSwift"),
         .external(name: "KeychainSwift"),
-        .external(name: "CombineExt"),
-        .external(name: "SwiftLinkPreview")
+        .external(name: "CombineExt")
     ]
 )
