+++ Workspace.swift
@@ -24,6 +24,7 @@ let workspace = Workspace(name: "TodoCalendar",
                             "Services/SpeechService",
                             "Services/PlaceService",
                             "Services/AuthService",
+                            "Services/ExternalServices",
                             "Supports/Extensions",
                             "Supports/UnitTestHelpKit",
                             "Supports/Common3rdParty",
